backend/requirements.txt
Django==5.1.2
djangorestframework
django-cors-headers
celery
ocrmypdf
ollama
langchain-ollama
PyPDF2
langchain-postgres
langchain
redis
celery[redis]==5.2.7
pdfminer.six
textblob
scikit-learn
numpy
semchunk
tiktoken
marker-pdf
marker
torch
torchvision
torchaudio
python-dotenv
psycopg2-binary>=2.8
markitdown
pyyaml==6.0.1
djangorestframework-simplejwt
google-auth
google-auth-oauthlib
langgraph